+++ application/tests/Bootstrap.php
@@ -319,7 +319,6 @@ define('VIEWPATH', $view_folder.DIRECTORY_SEPARATOR);
  * If you want to use monkey patching, uncomment below code and configure
  * for your application.
  */
-/*
 require __DIR__ . '/_ci_phpunit_test/patcher/bootstrap.php';
 MonkeyPatchManager::init([
 	// PHP Parser: PREFER_PHP7, PREFER_PHP5, ONLY_PHP7, ONLY_PHP5
@@ -346,11 +345,11 @@ MonkeyPatchManager::init([
 	],
 	// Additional functions to patch
 	'functions_to_patch' => [
+		'log_message'
 		//'random_string',
 	],
 	'exit_exception_classname' => 'CIPHPUnitTestExitException',
 ]);
-*/
 
 /*
  * -------------------------------------------------------------------