+++ application/tests/models/Sites_Model_test.php
@@ -23,6 +23,13 @@ class Site_Model_test extends TestCase {
 		$this->assertRegExp('/^[c|v][0-9\.]+(?:\/c[0-9\.]+)?$/', $result['latest_chapter']);
 		$this->assertRegExp('/^[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+:[0-9]+$/', $result['last_updated']);
 	}
+	public function test_MangaFox_fail() {
+		MonkeyPatch::patchFunction('log_message', NULL, 'MangaFox'); //Stop logging stuff...
+		$result = $this->Sites_Model->{'MangaFox'}->getTitleData('i_am_a_bad_url');
+
+		$this->assertNull($result);
+		MonkeyPatch::verifyInvokedOnce('log_message', ['error', 'Series missing? (MangaFox): i_am_a_bad_url']);
+	}
 
 	public function test_MangaHere() {
 		$result = $this->Sites_Model->{'MangaHere'}->getTitleData('tsugumomo');